\name{HEIfunctions-package}
\alias{HEIfunctions-package}
\alias{HEIfunctions}
\docType{package}
\title{
HEI grant functions
}
\description{
Functions for internal use related to HEI grant.
}
\details{
\tabular{ll}{
Package: \tab HEIfunctions\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-05\cr
License: \tab GPL-2 \cr
}
}
\author{
Lauren

Maintainer: Lauren <lbhund@gmail.com>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
