\name{pm10west}
\alias{pm10west}
\docType{data}
\title{
PM10 in the West
}
\description{
PM10 in the West
}
\usage{data(pm10west)}
\format{
  A data frame with 711 observations on the following 44 variables.
  \describe{
    \item{\code{Monitor_ID}}{a numeric vector}
    \item{\code{Latitude}}{a numeric vector}
    \item{\code{Longitude}}{a numeric vector}
    \item{\code{Land.Use}}{a character vector}
    \item{\code{Location.Setting}}{a character vector}
    \item{\code{a}}{a numeric vector}
    \item{\code{Median_income}}{a numeric vector}
    \item{\code{HS_rate}}{a numeric vector}
    \item{\code{Urban_rate}}{a numeric vector}
    \item{\code{Migration_5_year_rate}}{a numeric vector}
    \item{\code{Hispanic_cens}}{a numeric vector}
    \item{\code{White_cens}}{a numeric vector}
    \item{\code{Black_cens}}{a numeric vector}
    \item{\code{Other_cens}}{a numeric vector}
    \item{\code{Current_Smoking_rate_weighted}}{a numeric vector}
    \item{\code{HOUSEUNIT}}{a numeric vector}
    \item{\code{POPULATION}}{a numeric vector}
    \item{\code{Female_cens}}{a numeric vector}
    \item{\code{Age_mean}}{a numeric vector}
    \item{\code{Female_rate.med}}{a numeric vector}
    \item{\code{White_rate.med}}{a numeric vector}
    \item{\code{Black_rate.med}}{a numeric vector}
    \item{\code{Tot_ALL_cause_death}}{a numeric vector}
    \item{\code{Person_years}}{a numeric vector}
    \item{\code{Total_den}}{a numeric vector}
    \item{\code{Respiratory_adm}}{a numeric vector}
    \item{\code{CVD_adm}}{a numeric vector}
    \item{\code{Dew_point_annual_mean}}{a numeric vector}
    \item{\code{temperature_annual_mean}}{a numeric vector}
    \item{\code{temperature_annual_max_mean}}{a numeric vector}
    \item{\code{temperature_annual_min_mean}}{a numeric vector}
    \item{\code{pm10base}}{a numeric vector}
    \item{\code{pm10fu}}{a numeric vector}
    \item{\code{weight_cdc}}{a numeric vector}
    \item{\code{smoking_cdc}}{a numeric vector}
    \item{\code{adulthouse_cdc}}{a numeric vector}
    \item{\code{exercise_cdc}}{a numeric vector}
    \item{\code{asthma_cdc}}{a numeric vector}
    \item{\code{HTN_cdc}}{a numeric vector}
    \item{\code{diabetes_cdc}}{a numeric vector}
    \item{\code{BMIover_cdc}}{a numeric vector}
    \item{\code{BMIob_cdc}}{a numeric vector}
    \item{\code{genhealth_cdc}}{a numeric vector}
    \item{\code{nodoc_cdc}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
}
\keyword{datasets}
