\name{psspatmodel}
\alias{psspatmodel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Spatial propensity score model
}
\description{
Spatial propensity score model using spBayes
}
\usage{
psspatmodel(formula, data, trt, coords, na.rm = T, ...)
}
\arguments{
  \item{formula}{model formula
}
  \item{data}{dataset
}
  \item{trt}{string variable with treatment indicator
}
  \item{coords}{coordinates
}
  \item{na.rm}{exclude observations with missing outcome?
}
  \item{...}{Other arguments to spLM}
}
\details{
}
\value{
}
\references{
}
\author{Lauren
}
\note{
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
