\name{spatpred}
\alias{spatpred}
\title{
Baseline spatial prediction
}
\description{
Predicts 
}
\usage{
spatpred(formula, data, data.p, coords, coords.p, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula}{model formula}
  \item{data}{orignial dataset}
  \item{data.p}{dataset for prediction}
  \item{coords}{original coordinates}
  \item{coords.p}{coordinates for prediction}
  \item{\dots}{other arguments to be passed to spLM}
}
\details{
}
\value{
}
\references{
}
\author{
Lauren
}
\note{
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
