\name{sum.miss}
\alias{sum.miss}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarize missingness.
}
\description{
Summarize missingness in covariates.
}
\usage{
sum.miss(whichvars, dat)
}
\arguments{
  \item{whichvars}{vector of strings containing variable names of interest}
  \item{dat}{data frame}
}
\details{
}
\value{
}
\references{
}
\author{
Lauren
}
\note{
}
\seealso{
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
