\name{summarizeps}
\alias{summarizeps}
\title{
Summarize propensity score
}
\description{
Summarize propensity score.
}
\usage{
summarizeps(ps, a, data, nextreme)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ps}{string with propensity score}
  \item{a}{string with intervention}
  \item{data}{dataset}
  \item{nextreme}{number of extreme values of PS to display}
}
\details{
}
\value{
}
\references{
}
\author{
Lauren
}
\note{
}
\seealso{
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
