\name{summary.psbayes}
\alias{summary.psbayes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary function for Bayesian propensity score regression.
}
\description{
Summarizes bayesian propensity score model.
}
\usage{
summary.psbayes(out, full = T)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{out}{
object of class psbayes.
}
  \item{full}{
if TRUE, prints summary from the two counterfactual models.
}
}
\details{
}
\value{
}
\references{
}
\author{
Lauren
}
\note{
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
