\name{summary.psregression}
\alias{summary.psregression}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarize object of class psregression.
}
\description{
Summary function for frequentist propensity score regression model.
}
\usage{
summary.psregression(out, full = T)
}
\arguments{
  \item{out}{
Object of class psregression.
}
  \item{full}{
if TRUE, prints summary from the two counterfactual models.
}
}
\details{
}
\value{
}
\references{
}
\author{
Lauren
}
\note{
}

\seealso{
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
