\name{summarypollution}
\alias{summarypollution}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarize pollution data
}
\description{
Summarizes the pollution data
}
\usage{
summarypollution(base, fu, a, data, na.rm = T)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{base}{string for baseline pollution}
  \item{fu}{string for follow up pollution}
  \item{a}{string for intervention}
  \item{data}{dataset}
  \item{na.rm}{remove missing data if T.}
}
\details{
}
\value{
}
\references{
}
\author{
Lauren
}
\note{
}
\seealso{
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
